+++ src/bookmark_reader/target_writer.rs
@@ -13,7 +13,7 @@ where
 {
     fn write(&mut self, target_bookmarks: &TargetBookmarks) -> Result<(), BogrepError> {
         let bookmarks = JsonBookmarks::from(target_bookmarks);
-        let json = json::serialize(bookmarks)?;
+        let json = json::serialize(&bookmarks)?;
 
         self.write_all(&json).map_err(BogrepError::WriteFile)?;