+++ src/client.rs
@@ -6,10 +6,7 @@ use headless_chrome::{Browser, LaunchOptionsBuilder};
 use log::{debug, error, trace};
 use parking_lot::Mutex;
 use reqwest::{
-    header::{
-        HeaderMap, HeaderName, HeaderValue, ACCEPT, ACCEPT_ENCODING, ACCEPT_LANGUAGE,
-        CACHE_CONTROL, CONNECTION, HOST, USER_AGENT,
-    },
+    header::{HeaderMap, HeaderValue, ACCEPT, ACCEPT_ENCODING, ACCEPT_LANGUAGE, HOST, USER_AGENT},
     Client as ReqwestClient, Url,
 };
 use std::{