+++ src/cmd/mod.rs
@@ -3,15 +3,17 @@ mod clean;
 mod configure;
 mod fetch;
 mod import;
+mod init;
 mod remove;
 mod search;
 mod sync;
 
 pub use add::add;
 pub use clean::clean;
-pub use configure::{configure, configure_sources};
+pub use configure::configure;
 pub use fetch::fetch;
 pub use import::import;
+pub use init::{init, init_sources};
 pub use remove::remove;
 pub use search::search;
 pub use sync::sync;